Dataset schema
A `dataset` defines the `form` of a population within pear. There
is no PBSHM data stored directly agaisnt the dataset object,
instead a dataset has a set of `state`s (starting at zero) which
snapshot the state of the population. Any actions on a `state`
that would cause modifications to the SI or IE models result in
a new `state` for the dataset. Any actions on a `state` that result
in generated simulations (channels/features) are a `scenario`. Any
change that would effect the `form` of the population is a new
`dataset`.